Correcting mistakes
* changed best value from int to double
* updating manually the best value in hillclimb best (because update
  method changes encoding to binary)

+++ HW1/ga/GeneticAlgorithm.h
@@ -77,8 +77,8 @@ class GeneticAlgorithm
     double evaluateChromozome(const chromozome& chromozome, std::size_t index);
     double evaluateChromozomeAndUpdateBest(std::size_t index);
 
-    void updateBestChromozome(int newValue, const chromozome& newBest);
-    void updateBestChromozome(int newValue, std::size_t index);
+    void updateBestChromozome(double newValue, const chromozome& newBest);
+    void updateBestChromozome(double newValue, std::size_t index);
 
     /// this has to be done sequentially
     void evaluatePopulation();
